Move completion streaming into `Thread`
